token-finding-service/requirements.txt
Flask==2.0.1
werkzeug==2.0.1
psycopg2-binary